fxtest: Reorganize into files
Break apart lifecycle and app into separate files with their own test
files.
